+++ src/main/java/com/drone/system/mapper/UavMapper.java
@@ -21,7 +21,7 @@ public interface UavMapper {
      * @param uavId 无人机ID
      * @return 无人机对象信息
      */
-    public Uav selectUavByUavId(Long uavId);
+    public Uav selectUavByUavId(Integer uavId);
     
     /**
      * 新增无人机
@@ -45,5 +45,5 @@ public interface UavMapper {
      * @param uavIds 需要删除的无人机ID
      * @return 结果
      */
-    public int deleteUavByUavIds(Long[] uavIds);
+    public int deleteUavByUavIds(Integer[] uavIds);
 }
